fix: relative path handling
By default if you pass a relative path to the .NET APIs, it will use [Environment]::CurrentDirectory as its base,
not your actual PowerShell working directory.

The problem is that the two are almost never the same.

This changes it so that the entry path is resolved by PowerShell, not .NET.